+++ spec/testdata/root1/mainRoot1/src/main.c
@@ -1,0 +1,5 @@
+extern int i;
+int main()
+{
+	return 0;
+}