+++ spec/testdata/root1/mainRoot1/src/main.cpp
@@ -1,0 +1,8 @@
+#include "main.h"
+#include "lib1.h"
+#include "lib2.h"
+
+int main()
+{
+	return 0;
+}